requirements.txt
tensorflow
pandas
scipy
numpy
matplotlib
absl-py